sizeof() yields a size_t, which may well be bigger than an int.  Use %lu as a format specifier for it, and cast it to unsigned long in case it's some other size.
These uses are just in diagnostic messages, so if size_t really is bigger
than u_long on a given platform, the only outcome will be that the error
message is misleading (and if any of these numbers are nearly that big,
something is very wrong anyway).

Monotone-Parent: d3626332051c78c70c31fcad8677cab12bdbedb6
Monotone-Revision: f83386668f89a08ca585808c5802bcc78e97879b

+++ list.c
@@ -118,8 +118,8 @@ void AddToList(name_list **list_head, char *name, char *ptr)
     if (nptr == NULL)
     {
 	twmrc_error_prefix();
-	fprintf (stderr, "unable to allocate %d bytes for name_list\n",
-		 sizeof(name_list));
+	fprintf (stderr, "unable to allocate %lu bytes for name_list\n",
+		 (unsigned long) sizeof(name_list));
 	Done(0);
     }
